Move nested class to a nested file (no type layout change made)
